fix double modal issue
